requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
pdfplumber
PyMuPDF
pytesseract
Pillow
python-dotenv
aiofiles
asyncio
pydantic
tiktoken
tqdm
streamlit
plotly
